ARM: dts: imx28-evk: add vmmc-supply for mmc device
Add vmmc-supply for mmc device, so that the power to mmc slot can be
turned on by mmc driver if bootloader hasn't done it.

Also remove "okay" status for mmc1/ssp1, as the device has pin conflict
with gpmi-nand.  Actually the device is already disabled since the
mxs-mmc driver adopts pinctrl support because there is no pinctrl setup
for the device in dts.  Make this explicit by removing "okay" status.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>

[ upstream commit: 64edbcdd9c8d2aac36bead7f8cdafe0ec02bab40 ]