ASoC: omap-mcbsp: Device tree binding documentation update
To reflect the final devicetree node structure of McBSPs.

The initial OMAP McBSP DT structure was not able to describe the IP (and
it's versions) correctly.  The main issue was the sidetone block of
McBSP2/3 on OMAP3.  With this change in the DT description the OS can
get the needed information about the IP.

The sidetone is still not supported when the Linux kernel is booted with DT
since we still depend on hwmod to fill the resources.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>

[ upstream commit: b8101048f0f3cd281ed4c4901e38ae2bcfb32030 ]