Merge tag 'for-usb-next-2012-09-13' of git://git.kernel.org/pub/scm/linux/kernel/git/sarah/xhci into usb-next
xHCI patches for 3.7

Hi Greg,

Here's seven patches for 3.7.

The first four fix an issue with Set Address command timeouts.  It turns out
that Set Address timeouts can trigger a warning that was put in to avoid a NULL
pointer dereference.  This patchset fixes the underlying cause of the NULL
pointer that was papered over by the warning.  They should be applied to stable,
but I'm a bit nervous about the size, so I'd rather they go into 3.7, rather
than trying to stuff them into a late 3.6-rc.

The other three patches are various trivial fixes.

Sarah Sharp

[ upstream commit: f80e866f721942bfd9e5134e17b75f8cf53ebf5c ]